requirements.txt
flask
pylette
colorthief
pillow
flask-login